requirements.txt
Quart~=0.19.5
marshmallow~=3.21.2
bcrypt~=4.1.3
tortoise-orm~=0.20.1
tortoise-orm[asyncmy]